backend/requirements.txt
Django==4.2
djangorestframework==3.15.1
djoser==2.2.0
psycopg2-binary==2.9.9
gunicorn==22.0.0
django-cors-headers==4.3.1
dj-database-url==2.2.0
python-dotenv==1.0.1